+++ index.php
@@ -59,7 +59,7 @@
                 <div class="col d-flex justify-content-end">
                     <div class="row">
                         <div class="col">
-                            <p class="text-center" id="version_text" style="font-weight: normal;padding-top: 12px;color: rgb(0,0,0);">CP 4.2.0a<br><a href="https://github.com/magdesign/PocketVJ-CP-exh/issues" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help_issue.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="issues"></a><a href="https://www.pocketvj.com/manual/index.html" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="read the f**king docs!"></a><br></p>
+                            <p class="text-center" id="version_text" style="font-weight: normal;padding-top: 12px;color: rgb(0,0,0);">CP 4.2.1<br><a href="https://github.com/magdesign/PocketVJ-CP-exh/issues" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help_issue.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="issues"></a><a href="https://www.pocketvj.com/manual/index.html" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="read the f**king docs!"></a><br></p>
                         </div>
                     </div>
                 </div>
@@ -809,9 +809,19 @@ elem2.addEventListener("input", range2Value);
                             </div>
                             <div class="col">
                                 <div>
-                                    <h3>&nbsp;</h3>
-                                    <div class="col colforbutton" style="height: 48px;"><a class="btn btn-primary action" role="button" data-toggle="tooltip" title="show webcam on screen with gender detection" href="?action=onscreenwebcamusbgender">Gender on Screen<br></a></div>
+                                    <h2>&nbsp;</h2>
+                                    <div class="col colforbutton"><a class="btn btn-primary action" role="button" data-toggle="tooltip" title="show webcam on screen with FX" href="?action=onscreenwebcamusbhsv">Show on Screen HSV<br></a></div>
                                     <div class="col colforbutton"><a class="btn btn-primary action" role="button" href="?action=stopwebcamusb" data-toggle="tooltip" title="turn remote webcam off">Webcam OFF<br></a></div>
+                                    <div class="col colforbutton">
+                                        <h3>&nbsp;</h3>
+                                    </div>
+                                </div>
+                            </div>
+                            <div class="col">
+                                <div>
+                                    <h3>&nbsp;</h3>
+                                    <div class="col colforbutton" style="height: 48px;"><a class="btn btn-primary action" role="button" data-toggle="tooltip" title="!!very experimental, read manual!!!! webcam with gender detection" href="?action=onscreenwebcamusbgender">Gender on Screen<br></a></div>
+                                    <div class="col colforbutton"></div>
                                 </div>
                             </div>
                             <div class="col">
@@ -2548,7 +2558,6 @@ elem2.addEventListener("input", range2Value);
     <script src="assets/js/jquery.min.js"></script>
     <script src="assets/bootstrap/js/bootstrap.min.js"></script>
 
-
 <!-- Sets the "Set Time" button function--> 
  
 <script type="text/javascript">
@@ -2611,7 +2620,6 @@ setInterval(run, 3000);
         });
     });
     </script>
-
 
 </body>